close #85

+++ app/Http/Controllers/API/ShowController.php
@@ -193,8 +193,10 @@ class ShowController extends Controller
         // show being joined. This request is therefore authorized and we can
         // now make the connection.
         $show->invitees()->detach($request->user()->id);
-        $show->hosts()->detach($request->user()->id);
-        $show->hosts()->attach($request->user()->id, ['accepted' => true]);
+        if (! $request->has('cancel')) {
+            $show->hosts()->detach($request->user()->id);
+            $show->hosts()->attach($request->user()->id, ['accepted' => true]);
+        }
 
         return $show;
     }
